requirements.txt
fastapi[standard]
sqlalchemy==2.0.39
python-jose[cryptography]
bcrypt==4.3.0